refactoring to relocate flask app in one create_app method which will be used for both prod and test envs
