+++ app/extensions.py
@@ -1,0 +1,5 @@
+from flask.ext.sqlalchemy import SQLAlchemy
+db = SQLAlchemy()
+
+from flask.ext.migrate import Migrate
+migrate = Migrate()
