+++ app/models/chunk.py
@@ -1,4 +1,4 @@
-from app import db
+from app.extensions import db
 
 
 class Chunk(db.Model):
